% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_printmd.R
\name{apa_corr}
\alias{apa_corr}
\title{Print correlation statistics in Markdown}
\usage{
apa_corr(x, digits = 2, pdigits = 3, pzero = FALSE, ci = TRUE)
}
\arguments{
\item{x}{Correlation object}

\item{digits}{Number of digits after the decimal for means, correlation
coefficient, confidence intervals, and degrees of freedom}

\item{pdigits}{Number of digits after the decimal for p-values (also controls
cutoff for small p-values)}

\item{pzero}{Logical indicator of whether to include leading zero for
p-values}

\item{ci}{Logical indicator of whether to print 95\% confidence intervals}
}
\value{
A character string of statistical information formatted in Markdown.
}
\description{
Print correlation statistics in Markdown
}
\examples{
# Prepare data
beavers <- merge(beaver1, beaver2, by = "time")
beavers <- beavers[, c("time", "temp.x", "temp.y")]
beavers_corr <- cor.test(beavers$temp.x, beavers$temp.y)
# Print statistics
apa_corr(beavers_corr)
# Change digits
apa_corr(beavers_corr, digits = 3)
# Change cutoff digits for p-value
apa_corr(beavers_corr, pdigits = 2)
# Add leading zero to p-value and don't print confidence intervals
apa_corr(beavers_corr, pzero = TRUE, ci = FALSE)
}
\seealso{
Other functions for printing statistics: 
\code{\link{apa_ttest}()}
}
\concept{functions for printing statistics}
