% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa_printmd.R
\name{apa_ttest}
\alias{apa_ttest}
\title{Print t-test statistics in Markdown}
\usage{
apa_ttest(x, digits = 1, pdigits = 3, pzero = FALSE, full = TRUE)
}
\arguments{
\item{x}{t-test object}

\item{digits}{Number of digits after the decimal for means, confidence
intervals, t-statistic, and degrees of freedom}

\item{pdigits}{Number of digits after the decimal for p-values (also controls
cutoff for small p-values)}

\item{pzero}{Logical indicator of whether to include leading zero for
p-values}

\item{full}{Logical indicator of whether to include means and confidence
intervals or just t-statistic and p-value}
}
\value{
A character string of statistical information formatted in Markdown.
}
\description{
Print t-test statistics in Markdown
}
\examples{
# Prepare data
beavers <- merge(beaver1, beaver2, by = "time")
beavers <- beavers[, c("time", "temp.x", "temp.y")]
beavers_tt <- t.test(beavers$temp.x, beavers$temp.y)
# Print statistics
apa_ttest(beavers_tt)
# Change digits
apa_ttest(beavers_tt, digits = 2)
# Change cutoff digits for p-value
apa_ttest(beavers_tt, pdigits = 2)
# Add leading zero to p-value and don't print confidence intervals
apa_ttest(beavers_tt, pzero = TRUE, full = FALSE)
}
\seealso{
Other functions for printing statistics: 
\code{\link{apa_corr}()}
}
\concept{functions for printing statistics}
